Move into root_checked.rs
