+++ src/hugr/views.rs
@@ -2,16 +2,16 @@
 
 pub mod descendants;
 pub mod petgraph;
+mod root_checked;
 pub mod sibling;
 pub mod sibling_subgraph;
 
 #[cfg(test)]
 mod tests;
 
-use std::marker::PhantomData;
-
 pub use self::petgraph::PetgraphWrapper;
 pub use descendants::DescendantsGraph;
+pub use root_checked::RootChecked;
 pub use sibling::SiblingGraph;
 pub use sibling_subgraph::SiblingSubgraph;
 
@@ -309,45 +309,6 @@ pub trait RootTagged: HugrView {
     type RootHandle: NodeHandle;
 }
 
-/// A view of the whole Hugr.
-/// (Just provides static checking of the type of the root node)
-pub struct RootChecked<H, Root = Node>(H, PhantomData<Root>);
-
-impl<H: HugrView, Root: NodeHandle> RootChecked<H, Root> {
-    /// Create a hierarchical view of a whole HUGR
-    ///
-    /// # Errors
-    /// Returns [`HugrError::InvalidTag`] if the root isn't a node of the required [`OpTag`]
-    pub fn try_new(hugr: H) -> Result<Self, HugrError> {
-        check_tag::<Root>(&hugr, hugr.root())?;
-        Ok(Self(hugr, PhantomData))
-    }
-}
-
-impl<Root> RootChecked<Hugr, Root> {
-    /// Extracts the underlying (owned) Hugr
-    pub fn into_hugr(self) -> Hugr {
-        self.0
-    }
-}
-
-impl<H: AsRef<Hugr>, Root: NodeHandle> RootTagged for RootChecked<H, Root> {
-    type RootHandle = Root;
-}
-
-impl<H: AsRef<Hugr>, Root> AsRef<Hugr> for RootChecked<H, Root> {
-    fn as_ref(&self) -> &Hugr {
-        self.0.as_ref()
-    }
-}
-
-impl<H: AsMut<Hugr> + AsRef<Hugr>, Root: NodeHandle> AsMut<Hugr> for RootChecked<H, Root> {
-    #[inline(always)]
-    fn as_mut(&mut self) -> &mut Hugr {
-        self.0.as_mut()
-    }
-}
-
 /// A common trait for views of a HUGR hierarchical subgraph.
 pub trait HierarchyView<'a>: RootTagged + Sized {
     /// Create a hierarchical view of a HUGR given a root node.
@@ -520,60 +481,5 @@ pub(crate) mod sealed {
         fn root_node(&self) -> Node {
             self.as_ref().root.into()
         }
-    }
-}
-
-#[cfg(test)]
-mod test {
-    use super::{NodeType, RootChecked};
-    use crate::extension::ExtensionSet;
-    use crate::hugr::hugrmut::sealed::HugrMutInternals;
-    use crate::hugr::{HugrError, HugrMut};
-    use crate::ops::handle::{CfgID, DataflowParentID, DfgID};
-    use crate::ops::{BasicBlock, LeafOp, OpTag};
-    use crate::{ops, type_row, types::FunctionType, Hugr, HugrView};
-
-    #[test]
-    fn root_checked() {
-        let root_type = NodeType::pure(ops::DFG {
-            signature: FunctionType::new(vec![], vec![]),
-        });
-        let mut h = Hugr::new(root_type.clone());
-        let cfg_v = RootChecked::<&Hugr, CfgID>::try_new(&h);
-        assert_eq!(
-            cfg_v.err(),
-            Some(HugrError::InvalidTag {
-                required: OpTag::Cfg,
-                actual: OpTag::Dfg
-            })
-        );
-        let mut dfg_v = RootChecked::<&mut Hugr, DfgID>::try_new(&mut h).unwrap();
-        // That is a HugrMutInternal, so we can try:
-        let root = dfg_v.root();
-        let bb = NodeType::pure(BasicBlock::DFB {
-            inputs: type_row![],
-            other_outputs: type_row![],
-            predicate_variants: vec![type_row![]],
-            extension_delta: ExtensionSet::new(),
-        });
-        let r = dfg_v.replace_op(root, bb.clone());
-        assert_eq!(
-            r,
-            Err(HugrError::InvalidTag {
-                required: OpTag::Dfg,
-                actual: ops::OpTag::BasicBlock
-            })
-        );
-        // That didn't do anything:
-        assert_eq!(dfg_v.get_nodetype(root), &root_type);
-
-        let mut dfp_v = RootChecked::<&mut Hugr, DataflowParentID>::try_new(&mut h).unwrap();
-        let r = dfp_v.replace_op(root, bb.clone());
-        assert_eq!(r, Ok(root_type));
-        assert_eq!(dfp_v.get_nodetype(root), &bb);
-
-        // And it's a HugrMut:
-        let nodetype = NodeType::pure(LeafOp::MakeTuple { tys: type_row![] });
-        dfp_v.add_node_with_parent(dfp_v.root(), nodetype).unwrap();
     }
 }
